(crudely) working shell
Takes in command from stdin and does a fork/exec of
the command. At the moment, no parsing of the commandline
is done, so only commands within the default system path
with no arguments will run successfully.

Also, none of the typical builtins for the shell are
implemented yet, e.g. `cd`